Cleanup commit
- moved AST code into its own file
- deleted a bunch of old working code I wanted in history
